# Don't copy buffer with file for each lookup