Before this commit lookup to maxmind db:
 - had synchronized call that was a bottle neck for multithread application;
 - copied bytebuffer for each lookup that made life of GC much worst.

Here I rewrited a bit decoder to make it thread safe and as result I removed duplicating of buffer.