@oschwald 2018-08-05
I think there may be a misunderstanding. `ByteBuffer.duplicate()` only creates a copy of the `ByteBuffer`. It does not create a copy of the underlying data. Do you have any profiling to show that this change significantly impacts the number of allocations? Similarly, as `duplicate()` is such a light operation (creating a tiny object) compared to everything else that the reader does, I am a bit skeptical that this is much of a win for multithreaded applications.

The `ByteBuffer` `get*()` methods are not documented as thread safe, which is why the `BufferHolder` was added in the first place.