@oschwald 2018-08-05
This also appears significantly slower than master for me using the benchmark in the repo.

Master:

```
No caching
Warming up
Requests per second: 133651
Requests per second: 155788
Requests per second: 146648

Benchmarking
Requests per second: 157624
Requests per second: 161356
Requests per second: 164758
Requests per second: 165471
Requests per second: 163246

With caching
Warming up
Requests per second: 672478
Requests per second: 820981
Requests per second: 819900

Benchmarking
Requests per second: 816492
Requests per second: 837111
Requests per second: 837186
Requests per second: 834656
Requests per second: 827388
```

This branch:

```
No caching
Warming up
Requests per second: 97795
Requests per second: 102460
Requests per second: 102178

Benchmarking
Requests per second: 101960
Requests per second: 101757
Requests per second: 100319
Requests per second: 100408
Requests per second: 100015

With caching
Warming up
Requests per second: 483335
Requests per second: 566564
Requests per second: 566079

Benchmarking
Requests per second: 581114
Requests per second: 578708
Requests per second: 569587
Requests per second: 571731
Requests per second: 584128
````