@catap 2018-08-05
@oschwald well, I made [my benchmark](https://gist.github.com/catap/4ae2cb5000463abef253486ee21d0485) over OpenJDK's JMH.

I run it as
```
java -jar benchmark/target/benchmarks.jar -t 1 -f 1 -tu us -bm avgt -prof gc -jvmArgs -DmaxmindFile=$(pwd)/GeoIP2-City.mmdb MaxmindBenchmark.city_16
```
on
```
JDK 9.0.4, Java HotSpot(TM) 64-Bit Server VM, 9.0.4+11
```

Short summary: my benchmark confirm that it is slower (30us vs 25us) but it uses less memory (1.2Gb/sec vs 0.9Gb/sec and costs 2 times less GC cycles)

About performance costs I'm on this right now.

And you mentored a good point about thread safing of `ByteBuffer.get(..)`. I expect that it is (because it is read without any side effects) but I haven't got any proof.

master
```
Benchmark                                                                        (mode)  Mode  Cnt      Score     Error   Units
MaxmindBenchmark.city_16                                                         MEMORY  avgt    5     25.248 ±   2.455   us/op
MaxmindBenchmark.city_16:·gc.alloc.rate                                          MEMORY  avgt    5   1139.718 ± 108.498  MB/sec
MaxmindBenchmark.city_16:·gc.alloc.rate.norm                                     MEMORY  avgt    5  31680.001 ±   0.001    B/op
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'              MEMORY  avgt    5      0.015 ±   0.019  MB/sec
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'.norm         MEMORY  avgt    5      0.416 ±   0.514    B/op
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen                                    MEMORY  avgt    5   1141.999 ± 112.903  MB/sec
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen.norm                               MEMORY  avgt    5  31742.894 ± 338.623    B/op
MaxmindBenchmark.city_16:·gc.count                                               MEMORY  avgt    5    822.000            counts
MaxmindBenchmark.city_16:·gc.time                                                MEMORY  avgt    5    616.000                ms
MaxmindBenchmark.city_16                                                  MEMORY_MAPPED  avgt    5     23.931 ±   2.005   us/op
MaxmindBenchmark.city_16:·gc.alloc.rate                                   MEMORY_MAPPED  avgt    5   1203.032 ±  98.991  MB/sec
MaxmindBenchmark.city_16:·gc.alloc.rate.norm                              MEMORY_MAPPED  avgt    5  31696.001 ±   0.001    B/op
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'       MEMORY_MAPPED  avgt    5      0.015 ±   0.002  MB/sec
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'.norm  MEMORY_MAPPED  avgt    5      0.384 ±   0.046    B/op
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen                             MEMORY_MAPPED  avgt    5   1206.010 ± 100.458  MB/sec
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen.norm                        MEMORY_MAPPED  avgt    5  31774.357 ± 272.890    B/op
MaxmindBenchmark.city_16:·gc.count                                        MEMORY_MAPPED  avgt    5    576.000            counts
MaxmindBenchmark.city_16:·gc.time                                         MEMORY_MAPPED  avgt    5    449.000                ms
```

this branch:
```
Benchmark                                                                        (mode)  Mode  Cnt      Score     Error   Units
MaxmindBenchmark.city_16                                                         MEMORY  avgt    5     29.200 ±   1.106   us/op
MaxmindBenchmark.city_16:·gc.alloc.rate                                          MEMORY  avgt    5    923.087 ±  35.062  MB/sec
MaxmindBenchmark.city_16:·gc.alloc.rate.norm                                     MEMORY  avgt    5  29688.001 ±   0.001    B/op
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'              MEMORY  avgt    5      0.009 ±   0.004  MB/sec
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'.norm         MEMORY  avgt    5      0.279 ±   0.144    B/op
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen                                    MEMORY  avgt    5    924.952 ±  39.864  MB/sec
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen.norm                               MEMORY  avgt    5  29747.756 ± 336.204    B/op
MaxmindBenchmark.city_16:·gc.count                                               MEMORY  avgt    5    450.000            counts
MaxmindBenchmark.city_16:·gc.time                                                MEMORY  avgt    5    327.000                ms
MaxmindBenchmark.city_16                                                  MEMORY_MAPPED  avgt    5     30.094 ±   4.262   us/op
MaxmindBenchmark.city_16:·gc.alloc.rate                                   MEMORY_MAPPED  avgt    5    894.222 ± 123.418  MB/sec
MaxmindBenchmark.city_16:·gc.alloc.rate.norm                              MEMORY_MAPPED  avgt    5  29608.001 ±   0.001    B/op
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'       MEMORY_MAPPED  avgt    5      0.006 ±   0.004  MB/sec
MaxmindBenchmark.city_16:·gc.churn.CodeHeap_'non-profiled_nmethods'.norm  MEMORY_MAPPED  avgt    5      0.209 ±   0.147    B/op
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen                             MEMORY_MAPPED  avgt    5    896.066 ± 122.754  MB/sec
MaxmindBenchmark.city_16:·gc.churn.G1_Old_Gen.norm                        MEMORY_MAPPED  avgt    5  29669.363 ± 257.249    B/op
MaxmindBenchmark.city_16:·gc.count                                        MEMORY_MAPPED  avgt    5    428.000            counts
MaxmindBenchmark.city_16:·gc.time                                         MEMORY_MAPPED  avgt    5    330.000                ms
```