@hc-codersatlas 2018-08-05
@oschwald are we sure the benchmark is correct? Might be a better idea to use JMH or similar. Unless TRACE is on nothing happens to the variable `t`. It can easily get escaped via escape analysis and the whole block disappearing. There could be other side effects. It also doesn't minimize what's getting tested - `random` and `inet` is part of the timing. These should be taken out, created first and stored in an arrray to bench this properly.